% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{allcredsetsPP}
\alias{allcredsetsPP}
\title{Construct a credible set for each trait and under each of single and multi-trait fine-mapping, and provide SNP PP details}
\usage{
allcredsetsPP(mpp.pp, cred = 0.99)
}
\arguments{
\item{mpp.pp}{object created in flashfm output, e.g. fm$mpp.pp if fm is output from FLASHFMwithJAM or FLASHFMwithFINEMAP}

\item{cred}{probability for credible set; default is 0.99}
}
\value{
list with 3 components: list of single-trait fine-mapping credible sets for each trait with the SNP PP for each variant, list of multi-trait fine-mapping credible sets for each trait with the SNP PP for each variant, credible set probability
}
\description{
Construct a credible set for each trait and under each of single and multi-trait fine-mapping, and provide SNP PP details
}
\author{
Jenn Asimit
}
