% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{calcABFo}
\alias{calcABFo}
\title{Calculate approximate Bayes' factor (ABF)}
\usage{
calcABFo(mod, mbeta, SSy, Sxy, Vy, N)
}
\arguments{
\item{mod}{joint SNP model with snps separated by \code{"\%"} e.g. \code{"snp1\%snp2"}}

\item{mbeta}{joint effect estimates for SNPs in model given by mod; output from multibeta}

\item{SSy}{sum(y.squared) for trait y}

\item{Sxy}{vector of sum(xy) for snp x, trait y}

\item{Vy}{trait variance}

\item{N}{sample size}
}
\value{
ABF for model mod
}
\description{
Calculate approximate Bayes' factor (ABF)
}
\author{
Jenn Asimit
}
