% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corX.list.R
\docType{data}
\name{corX.list}
\alias{corX.list}
\title{List of SNP correlation matrices for two groups}
\format{
An object of class \code{list} of length 2.
}
\usage{
corX.list
}
\description{
The first SNP correlation matrix in the list is for Group 1 and is 201 x 201, and the second matrix is for Group 2 with dimensions 329 x 329.
For a region in chromosome 19:45386029-45439498 (GRCh37/hg19), containing APOE, we used HapGen2 to generate two groups of individuals based on cohorts from the 1000 Genomes Project:
N1=90,000 based on CEU+TSI and N2=10,000 based on YRI+LWK.
Within each group, we retained variants with MAF > 0.001, regardless of whether the variant satisfies our MAF threshold in the other group.
This resulted in 201 variants in Group 1 and 329 variants in Group 2; 132 variants appear in both groups and there are 398 unique variants present in at least one of the groups.
}
\keyword{datasets}
