% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvs.R
\docType{data}
\name{cvs}
\alias{cvs}
\title{Vector of rsids for causal variants of two simulated traits in two groups}
\format{
An object of class \code{character} of length 3.
}
\usage{
cvs
}
\description{
This is the vector c(c1,c2,c3) = c("chr19_45387034_A_C", "chr19_45425175_T_G", "chr19_45427353_C_G") for the provided simulation example of two traits
where one trait has causal variants c1+c2 and the other trait has causal variants c1+c3; the traits have shared causal variant
c1. Trait 1 has the same causal variants in Groups 1 and 2, and likewise for Trait 2.
}
\keyword{datasets}
