% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{FLASHFMwithJAMd}
\alias{FLASHFMwithJAMd}
\title{Wrapper for flashfm Multi-Trait Fine-Mapping with JAM  - this is the dynamic number of max causal variant version}
\usage{
FLASHFMwithJAMd(
  gwas.list,
  corX,
  ybar,
  N,
  save.path,
  TOdds = 1,
  covY,
  cpp = 0.99,
  NCORES,
  maxcv = 1,
  maxcv_stop = 20,
  maxcv_autocheck = TRUE,
  jam.nM.iter = 1
)
}
\arguments{
\item{gwas.list}{List of M data.frame objects, where M is the number of traits (at most 6); gwas.list[[i]] is a data.frame for  trait i with 3 columns named: rsID, beta, EAF}

\item{corX}{SNP correlation matrix}

\item{ybar}{trait mean; if trait is transformed to be standard Normal, set ybar = 0, which is default}

\item{N}{Vector of length M; Nall[i] is the (effective) sample size for  trait i}

\item{save.path}{Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{TOdds}{target odds of no sharing to sharing; default is 1}

\item{covY}{trait covariance matrix (for at most 5 traits and all traits should have a signal in the region, e.g. min p < 1E-6)}

\item{cpp}{cumulative posterior probability threshold for selecting top models; default 0.99}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=A, but if on Windows, use NCORES=1}

\item{maxcv}{starting value for maximum number of causal variants; default 1}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants}

\item{maxcv_autocheck}{Logical for whether to incrementally increase maxcv (TRUE); if do not want to increment maxcv set this to FALSE}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; defailt 1 (1 million)}
}
\value{
list with 2 components: mpp.pp, a list with 4 components giving the SNP-level results (mpp.pp$PP,mpp.pp$MPP) and SNP group level results (mpp.pp$MPPg, mpp.pp$PPg); and snpGroups,
a list with 2 components giving the SNP groups construced under single-trait (snpGroups[[1]]) and multi-trait fine-mapping (snpGroups[[2]])
}
\description{
Wrapper for flashfm Multi-Trait Fine-Mapping with JAM  - this is the dynamic number of max causal variant version
}
\author{
Jenn Asimit
}
