% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g2t1CS.R
\docType{data}
\name{g2t1CS}
\alias{g2t1CS}
\title{Output from running JAMdynamic on Group 2, Trait 1 simulated dataset in this package}
\format{
An object of class \code{list} of length 3.
}
\usage{
g2t1CS
}
\description{
This is a list with three components:
g2t1CS$CS: the variants in the CS99, together with their marginal posterior probability of being causal (MPP)
g2t1CS$MPP: the MPP for all variants with non-negligible MPP
g2t1CS$PP: the models with non-negligible model PP
}
\keyword{datasets}
