% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups2update.R
\name{groupmultiU}
\alias{groupmultiU}
\title{Group SNPs; adapted from group.multi by Chris Wallace}
\usage{
groupmultiU(
  SM2,
  snp.data,
  is.snpmat,
  min.mppi = 0.01,
  minsnpmppi = 0.01,
  r2.minmerge = 0.6
)
}
\arguments{
\item{SM2}{snpmod object}

\item{snp.data}{SnpMatrix or SNP covariance matrix with snp ids as column names}

\item{is.snpmat}{logical taking value TRUE when SnpMatrix is provided and FALSE when covariance matrix is given}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all
diseases}

\item{minsnpmppi}{only group snps with total MPPI > minsnpmppi}

\item{r2.minmerge}{merge groups with minimum between-group r2 >
r2.minmerge}
}
\value{
list with three components.

First is a data.frame with each row giving summary statistics for
each group.

Second is a groups object, each elements ordered according to the rows of the summary

Third is the r2 matrix calculated.
}
\description{
Group SNPs; adapted from group.multi by Chris Wallace
}
