% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{JAMdynamic}
\alias{JAMdynamic}
\title{Expanded version of JAM (a single-trait fine-mapping approach) that first runs on thinned SNPs and then expands models on tag SNPs
This version starts at a low upper bound for max causal variants and decides on max upper bound based on data}
\usage{
JAMdynamic(
  gwas,
  corX,
  ybar = 0,
  Vy = 1,
  N,
  cred = 0.99,
  save.path,
  maxcv = 1,
  maxcv_stop = 20,
  jam.nM.iter = 5,
  r2 = 0.8,
  extra.java.arguments = NULL
)
}
\arguments{
\item{gwas}{data.frame with 3 columns named: rsID, beta, EAF}

\item{corX}{genotype correlation matrix (reference or from sample)}

\item{ybar}{trait mean; if trait is transformed to be standard Normal, set ybar = 0, which is default}

\item{Vy}{trait variance; if trait is transformed to be standard Normal, set Vy = 1, which is default}

\item{N}{sample size for trait; recommended to give effective sample sizes using GWAS summary statistics in Neff function}

\item{cred}{probability for credible set; default is 0.99}

\item{save.path}{path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{maxcv}{starting value for maximum number of causal variants}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants; maxcv_stop >= maxcv.}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; defailt 1 (1 million)}

\item{r2}{r.squared threshold for to find tag SNPs to reduce model search space before expanding to all SNPs;}

\item{extra.java.arguments}{A character string to be passed through to the java command line. E.g. to specify a
different temporary directory by passing "-Djava.io.tmpdir=/Temp".}
}
\value{
List of credible set variants with their MPP, MPP for all variants, PP for all models
}
\description{
Expanded version of JAM (a single-trait fine-mapping approach) that first runs on thinned SNPs and then expands models on tag SNPs
This version starts at a low upper bound for max causal variants and decides on max upper bound based on data
}
\author{
Feng Zhou
}
