% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprep.R
\name{LDqc}
\alias{LDqc}
\title{Quality control of dosage genotype matrix prior to LD calculation}
\usage{
LDqc(g, theta = 0.2, BestGuess = TRUE)
}
\arguments{
\item{g}{NxM genotype score matrix for N individuals and M variants}

\item{theta}{tolerance for dosage, e.g. if theta=0.2, then only dosages d that satisfy d<0.2, 0.8<d<1.2, and d>1.8 are retained}

\item{BestGuess}{if TRUE, then return "best guess" genotype matrix (default); if FALSE, return dosage genotype matrix, where dosages that do not meet
tolerance theta, are set to "NA"}
}
\value{
Outputs the "best guess" genotype matrix (if BestGuess=TRUE; default) or the "quality controlled" genotype matrix (if BestGuess=FALSE)
}
\description{
Quality control of dosage genotype matrix prior to LD calculation
}
\author{
Jenn Asimit
}
