% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groups2update.R
\name{makeSNPgroups2U}
\alias{makeSNPgroups2U}
\title{Make two sets of SNP groups using fine-mapping information from all of the traits using two sets of results and maps the names between them}
\usage{
makeSNPgroups2U(
  main.input,
  fm.multi,
  is.snpmat,
  min.mppi = 0.01,
  minsnpmppi = 0.001,
  r2.minmerge = 0.5
)
}
\arguments{
\item{main.input}{output from flashfm.input function}

\item{fm.multi}{output from flashfm function}

\item{is.snpmat}{logical taking value TRUE when genotype matrix is provided and FALSE when covariance matrix is given}

\item{min.mppi}{trim snp groups with total MPPI < min.mppi in all diseases; default 0.01}

\item{minsnpmppi}{only group snps with total MPPI > minsnpmppi; default 0.001}

\item{r2.minmerge}{merge groups with minimum between-group r2 > r2.minmerge; default 0.5}
}
\value{
list of  three objects: groups.fm is a list of SNP groups using the single-trait results; groups.flashfm is a list of SNP groups using the flashfm results; group.sizes is  a table of SNP group sizes for the two sets of groups
}
\description{
Make two sets of SNP groups using fine-mapping information from all of the traits using two sets of results and maps the names between them
}
