% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGflashfm.R
\name{MGflashfm}
\alias{MGflashfm}
\title{Multi-group  multi-trait fine-mapping, using output from FLASHFMwithJAM or FLASHFMwithFINEMAP; gives both multi- and single-trait results from flashfm object}
\usage{
MGflashfm(
  allflashfmPP,
  Nall,
  snps,
  cred = 0.99,
  multi = TRUE,
  cpp = 0.99,
  NCORES = 1
)
}
\arguments{
\item{allflashfmPP}{List of components with same length as number of Studies: allflashfmPP[[i]] = flashfmOUT[[i]]$mpp.pp for group  i}

\item{Nall}{List of components with same length as number of Studies: Nall[[i]] is the M-vector of trait sample sizes for group  i, where M is the number of traits}

\item{snps}{vector of all variants that exist in at least one group}

\item{cred}{Level for credible set; default 0.99}

\item{multi}{TRUE for multi-group  multi-trait fine-mapping; FALSE for multi-group  single-trait fine-mapping; default TRUE}

\item{cpp}{cumulative posterior probability threshold for selecting top models; this is ignored when maxmod is spespecified}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M, but if on Windows, use NCORES=1}
}
\value{
List consisting of two objects: summary = a list of multi-group  credible sets (variants and their msMPP), one for each trait; details = for each trait, list of all multi-group  models and variants and their msPP, msMPP
}
\description{
Multi-group  multi-trait fine-mapping, using output from FLASHFMwithJAM or FLASHFMwithFINEMAP; gives both multi- and single-trait results from flashfm object
}
\author{
Jenn Asimit
}
