% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGflashfm.R
\name{MGflashfmRET}
\alias{MGflashfmRET}
\title{Multi-group  multi-trait fine-mapping, using output from FLASHFMwithJAM or FLASHFMwithFINEMAP; gives both multi- and single-trait results from flashfm object}
\usage{
MGflashfmRET(
  gwas.list,
  flashfm.list,
  Nall,
  cred = 0.99,
  multi = FALSE,
  cpp = 0.99,
  NCORES = 1
)
}
\arguments{
\item{gwas.list}{List of A lists objects, where A is the number of groups; gwas.list[[i]] is a list for group  i with M data.frames (one for each trait)
with 3 columns named: rsID, beta, EAF;
if trait names are provided for the M data.frames (same names across groups), these trait names are given in output
if group  names are provided for the A data.frames, these group  names are given in the output}

\item{flashfm.list}{List of flashfm output from each of the groups}

\item{Nall}{List of components with same length as number of groups: Nall[[i]] is the M-vector of trait sample sizes for group  i, where M is the number of traits}

\item{cred}{Level for credible set; default 0.99}

\item{multi}{TRUE for multi-group  multi-trait fine-mapping; FALSE for multi-group  single-trait fine-mapping; default TRUE}

\item{cpp}{cumulative posterior probability threshold for selecting top models; this is ignored when maxmod is spespecified}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=M, but if on Windows, use NCORES=1}
}
\value{
List consisting of two objects: CSsummary = List of one data.frame for each trait; each trait data.frame gives the variants in the multi-group  credible set for the trait, the msMPP, pooled MAF, proportion of groups
that contain the variant, names of groups that contain the variant
CSdetail = [[1]] a list of multi-group  credible sets (variants and their msMPP), one for each trait; [[2]] details = for each trait, list of all multi-group  models and variants and their msPP, msMPP
}
\description{
Multi-group  multi-trait fine-mapping, using output from FLASHFMwithJAM or FLASHFMwithFINEMAP; gives both multi- and single-trait results from flashfm object
}
\author{
Jenn Asimit
}
