% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGfm.R
\name{MGfm}
\alias{MGfm}
\title{Multi-Group Single-Trait Fine-Mapping credible sets}
\usage{
MGfm(stfm.list, isnps, N, cred = 0.99, cpp = 0.99)
}
\arguments{
\item{stfm.list}{List of data.frame objects giving model PP from single-trait fine-mapping in each group}

\item{isnps}{Vector of variants that are present in at least one group's GWAS}

\item{N}{Vector of sample sizes in same order as groups in stfm.list}

\item{cred}{Level for credible set, default 0.99}

\item{cpp}{cumulative posterior probability threshold for selecting top models; default 0.99}
}
\value{
List of three objects: summary: data.frame of variants that belong to the credible set and their multi-group MPP (mgMPP), details: list of mgMPP for all variants and mgPP for the top multi-group models, cred:credible set level
}
\description{
Multi-Group Single-Trait Fine-Mapping credible sets
}
\author{
Jenn Asimit
}
