% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGfm.R
\name{MGFMwithJAM}
\alias{MGFMwithJAM}
\title{Wrapper for credible sets from Multi-Group Single-Trait Fine-Mapping with JAM}
\usage{
MGFMwithJAM(
  gwas.list,
  corX.list,
  Nall,
  save.path,
  cpp = 0.99,
  cred = 0.99,
  maxcv = 1,
  maxcv_stop = 20,
  maxcv_autocheck = TRUE,
  NCORES = 1,
  jam.nM.iter = 1
)
}
\arguments{
\item{gwas.list}{List of A data.frame objects, where A is the number of groups; gwas.list[[i]] is a data.frame for group i with 3 columns named: rsID, beta, EAF}

\item{corX.list}{List of A data.frame objects, where A is the number of groups; corX.list[[i]] is the SNP correlation matrix for group i}

\item{Nall}{Vector of length A; Nall[i] is the (effective) sample size for group i}

\item{save.path}{Path to save JAM output files; tmp files and could delete these later e.g. save.path=paste0(DIRout,"/tmpJAM/region1").}

\item{cpp}{cumulative posterior probability threshold for selecting top models; default 0.99}

\item{cred}{Level for credible set, default 0.99}

\item{maxcv}{starting value for maximum number of causal variants}

\item{maxcv_stop}{maximum value to consider for maximum number of causal variants}

\item{maxcv_autocheck}{Logical for whether to incrementally increase maxcv (TRUE); if do not want to increment maxcv set this to FALSE}

\item{NCORES}{number of cores for parallel computing; recommend NCORES=A, but if on Windows, use NCORES=1}

\item{jam.nM.iter}{in millions, number of iterations to use in JAM; defailt 1 (1 million)}
}
\value{
List consisting of two objects: CSsummary = List of one data.frame for each trait; each trait data.frame gives the variants in the multi-group  credible set for the trait, the mgMPP, pooled MAF, proportion of studies
that contain the variant, names of studies that contain the variant
CSdetail = [[1]] a list of multi-group  credible sets (variants and their mgMPP), one for each trait; [[2]] details = for each trait, list of all multi-group  models and variants and their mgPP, mgMPP
}
\description{
Wrapper for credible sets from Multi-Group Single-Trait Fine-Mapping with JAM
}
\author{
Jenn Asimit
}
