% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgmtCS.R
\docType{data}
\name{mgmtCS}
\alias{mgmtCS}
\title{Output from running MGFLASHFMwithJAM on the simulated dataset in this package}
\format{
An object of class \code{list} of length 3.
}
\usage{
mgmtCS
}
\description{
This is a list with three components:
mgmtCS$CSsummary is a list with 2 components giving the multi-group (multi-trait adjusted) 99\% credible sets for the two traits, including variant information
mgmtCS$CSdetail is a list that gives, for each trait, details on the multi-group (multi-trait adjusted) marginal PP of a SNP being causal (mgMPP) and the model PP (mgPP)
mgmtCS$flashfm.out gives the flashfm output for each group
}
\keyword{datasets}
