% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{multibetao}
\alias{multibetao}
\title{Using summary statistics, calculates joint effect estimates}
\usage{
multibetao(
  mod,
  beta1,
  Gmat,
  N,
  ybar,
  is.snpmat,
  raf = NULL,
  high_beta_thresh = 5
)
}
\arguments{
\item{mod}{joint SNP model with snps separated by \code{"\%"} e.g. \code{"snp1\%snp2"}}

\item{beta1}{named vector single-SNP effect estimates; each effect estimate has name given by SNP id and ids must appear in Gmat columns}

\item{Gmat}{genotype matrix with SNP columns and individuals rows; could be reference matrix or from sample}

\item{N}{sample size for trait}

\item{ybar}{trait mean}

\item{is.snpmat}{logical taking value TRUE when Gmat is a genotype matrix and FALSE when Gmat is a SNP covariance matrix}

\item{raf}{named vector of SNP reference allele frequencies (must match SNP coding used for effect estimates); only needed if Gmat is a covariance matrix}

\item{high_beta_thresh}{if absolute value of effect estimate < nigh_beta_thresh, the model is re-fit after checking for and removing variants to avoid high correlation}
}
\value{
joint effect estimates for SNPs in model given by mod
}
\description{
Using summary statistics, calculates joint effect estimates
}
\author{
Jenn Asimit
}
