% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nall.R
\docType{data}
\name{Nall}
\alias{Nall}
\title{List of samples size vectors for two simulated traits in two groups}
\format{
An object of class \code{list} of length 2.
}
\usage{
Nall
}
\description{
This is a list of two vectors: trait 1, 2 sample sizes for group 1, c(90000,90000) and trait 1,2 sample sizes for group 2, c(10000,10000)
for the provided simulation example.
}
\keyword{datasets}
