% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flashfm-updated.R
\name{Neff}
\alias{Neff}
\title{Approximate effective sample size from GWAS of related samples (copied from flashfm)}
\usage{
Neff(raf, seB, Vy = 1)
}
\arguments{
\item{raf}{vector of reference (or minor) allele frequencies for all SNPs in the GWAS}

\item{seB}{vector of standard errors of SNP effect estimates, in same order as SNPs in raf}

\item{Vy}{trait variance; default is 1 under assumption trait is transformed to a standard Normal distribution}
}
\value{
Estimate of effective sample size
}
\description{
Approximate effective sample size from GWAS of related samples (copied from flashfm)
}
\author{
Jenn Asimit
}
